Moved FileStream object to the restclient.entities package since it's not a resource object.
